requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
colorama==0.4.6
distro==1.9.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.29.0
playwright==1.58.0
psutil==6.1.1
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pyee==13.0.1
pygments==2.19.2
python-dotenv==1.2.2
pyyaml==6.0.3
rich==14.3.3
sniffio==1.3.1
taskipy==1.14.1
tomli==2.4.1
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
